requirements/requirements-gpu.txt
tensorflow-gpu==1.3.0